pyproject.toml
[project]
name = "repl-nix-bulkurlextractor"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "trafilatura>=1.12.2",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "markdown>=3.7",
    "aiohttp>=3.10.10",
    "asyncio>=3.4.3",
    "fastapi>=0.115.4",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.2",
    "codaio>=0.6.12",
    "werkzeug",
]
